Changed variable types to use less memory

+++ BasicOperationPlus/src/Sensors/Distance/DistanceSens.h
@@ -18,23 +18,23 @@ class DistanceSens{
                   The amount of sensors being used;
                   The pins connected to the "ECHO" pin on the HCSR04;
                   A timer to pause execution that functions as a yield to the processor */
-        DistanceSens(int trigPin, int sensorCount, uint8_t *sensorPins, unsigned long trigTimer);
+        DistanceSens(uint8_t trigPin, uint8_t sensorCount, uint8_t *sensorPins, uint16_t trigTimer);
         ~DistanceSens(); /* Class destructor */
       
         void start(void); /* Function starts the connection to the sensors. Best utilized as a thread */
 
         /* Function to set a different timer to the one previously set
            INPUT: The number of milliseconds between each execution */
-        void setTimer(unsigned long);
+        void setTimer(uint16_t);
 
         /* Specialized functions */
         double getDistanceL(void); /* Gets first result, that is the sensor on the left side of the car     */ 
         double getDistanceC(void); /* Gets second result, that is the sensor on the center side of the car  */
         double getDistanceR(void); /* Gets third result, that is the sensor on the right side of the car    */
     private:
-        int sensorCount;                /* Amount of sensors being used at once */
+        uint8_t sensorCount;                /* Amount of sensors being used at once */
         std::mutex mtx;                 /* Synchronization variable                                         */
-        unsigned long trigTimer;        /* Time between the executions of the "start" function              */
+        uint16_t trigTimer;        /* Time between the executions of the "start" function              */
         std::list<double *> distances;  /* List holding the results of the readings,                        */
                                         /* where a reading is made up of the distances read by each sensor  */
         double *averages;               /* Average between the previous FILTER_SAMPLE_NUM readings          */
